Merged PR 52: Add streaming support for s3storage
This adds support for all streaming modes with s3storage. It does this by implementing the StageAndCommit function which uses multipart uploads to write blocks that are used when doing streaming and by implementing GetFileBlockOffset which creates a set of blocks from an object of a given size. These are not the actual part sizes since the api that minio implements does not allow us to see that easily, but this fake list of blocks does allow all of streaming to work correctly.

When writing using streaming, it writes in blocks based on the part-size-mb config flag. There is one edge case is when writing new data to an existing object that is unaddressed. If the last part of the existing object is less than the minimum part size for AWS, then this block will be combined will the next block until they are larger than the minimum of 5MB as AWS requires all but the last block to be over 5MB but the last block can be any size. In very extreme edge cases this could cause a part to get larger than the maximum by being combined many times and exceed the maximum part size of 5GB, but this is a very extreme edge case and the solution would require us to redownload and resize the parts for the entire object to fix it, so I did not address this edge case.

I added many of the remaining tests in block_block_test and implemented those here that deal with StageAndCommit and GetFileBlockOffset. I also added support for updating the part-size-mb with the update config function.