Merged PR 54: Run S3 tests in Ci/CD pipeline
This adds support to run the s3 tests in our CI/CD pipeline. This runs on Lyve Cloud and runs the full unit and end to end tests for S3. The only remaining tests to run are the streaming tests, which can be added once the streaming pull request is merged.