Merged PR 64: Set the max input to int32 max
Qt has a built in automatic maximum of 99 in the absence of a set maximum input for a QSpinBox. In order to avoid such a low number, I set the absolute maximum number QSpinBox can take, which is a signed int32 (0x7FFFFFFF or 2,147,483,647 base ten). If there is such a time where Fuse can take a larger number than the GUI can provide, that element will have to change to accommodate a larger number, such as a QLineEdit with filtering.