Merged PR 65: Update Documentation
This fixes some renaming for cloudfuse in the markdown files and adds some information about building and installing on Windows as well as support for S3 storage. I updated the code of conduct from the Microsoft one to one that is used in other Seagate repositories and updated the security policy to the Seagate open source security policy.