Merged PR 72: Update blobfuse2.1.0
This pulls in the latest changes from the blobfuse 2.1.0 release. This adds a new block_cache component that I have changed to only work on Linux, since it is very specific to Linux. See https://github.com/Azure/azure-storage-fuse/releases/tag/blobfuse2-2.1.0 for the other release notes of the release.